requirements.txt
Flask
gunicorn
openai
fpdf2